MemoEntity를 Room 패키지로 이동
