--- app/src/main/java/com/babysloth/memo/data/model/MemoEntity.kt
@@ -1,19 +1,0 @@
-package com.babysloth.memo.data.model
-
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "memo")
-data class MemoEntity(
-    @PrimaryKey(autoGenerate = true)
-    val id: Long = 0,
-    @ColumnInfo(name = "create_at")
-    val createdAt: Long,
-    @ColumnInfo(name = "updated_at")
-    val updatedAt: Long,
-    val title: String,
-    val description: String,
-    @ColumnInfo(name = "is_bookmark")
-    val isBookmark: Boolean = false
-)
